C++: move the window related stuff in its own header
slint.h starts to be quite big, and lots of it is because of Window, so
moce Window to its own header
